\name{farva-package}
\alias{farva-package}
\alias{farva}
\docType{package}
\title{
\packageTitle{farva}
}
\description{
\packageDescription{farva}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{farva}
\packageIndices{farva}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{farva}

Maintainer: \packageMaintainer{farva}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
